Refactor : 연료 데이터를 레벨별 리스트 구조로 변경

+++ Assets/Scripts/SaveLoad/Data/SaveData.cs
@@ -36,7 +36,7 @@ public class PlayerSaveData
 public class CampFireSaveData
 {
     public int currentLevel;
-    public int currentFuel;
+    public float currentFuel;
     public bool isBurning;
     public float decreaseTimer;
 }
